fix: fix the issue of generating incorrect Sent time for message body
Description:

This commit fixes a bug in go-rsmq that caused the Sent time of message body to be generated incorrectly.
The original code used time.UnixMilli(sent) to convert the Sent timestamp from microseconds to milliseconds,
which caused the time to be incorrect.
This commit corrects the conversion to use time.UnixMicro(sent) instead,
which generates the correct Sent time.